requirements.txt
appdirs==1.4.4
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
docker-pycreds==0.4.0
filelock==3.12.0
fonttools==4.39.4
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
imageio==2.28.1
importlib-resources==5.12.0
Jinja2==3.1.2
kiwisolver==1.4.4
lazy_loader==0.2
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
packaging==23.1
pathtools==0.1.2
Pillow==9.5.0
protobuf==4.22.3
psutil==5.9.5
pycodestyle==2.10.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0
requests==2.29.0
scikit-image==0.20.0
scipy==1.9.1
sentry-sdk==1.21.1
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sympy==1.11.1
tifffile==2023.4.12
tomli==2.0.1
torch==2.0.0
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
vizdoom==1.1.14
wandb==0.15.1
zipp==3.15.0
